Add files via upload
I added this file to the master branch to show what I was typing from the terminal in hopes for some partial credit. I was getting an error when trying to do the final merge and did not know how to get rid of it. I hope that you can take a look and give some pointers and insight on how to fix the issue(s).